requirements.txt
fastapi
uvicorn
sentence-transformers
transformers
torch
faiss-cpu
pypdf
numpy
